Move this test to the MaxmindIpDataLookupsTests
